requirements.txt
# Core Dependencies for Streamlit Cloud Deployment
streamlit>=1.25.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0

# NLP & ML
scikit-learn>=1.3.0
nltk>=3.8.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# API (for local development)
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
